Create a module for app data handling
